+++ Package.swift
@@ -15,6 +15,9 @@ let package = Package(
             name: "FWApplicationCompatible",
             targets: ["FWApplication", "FWApplicationCompatible"]),
         .library(
+            name: "FWApplicationSwiftUI",
+            targets: ["FWApplication", "FWApplicationSwiftUI"]),
+        .library(
             name: "FWApplicationSDWebImage",
             targets: ["FWApplication", "FWApplicationSDWebImage"]),
         .library(
@@ -22,7 +25,7 @@ let package = Package(
             targets: ["FWApplication", "FWApplicationLottie"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/lszzy/FWFramework.git", from: "3.6.0"),
+        .package(url: "https://github.com/lszzy/FWFramework.git", from: "3.7.0"),
         .package(url: "https://github.com/SDWebImage/SDWebImage.git", from: "5.9.0"),
         .package(url: "https://github.com/airbnb/lottie-ios.git", from: "3.2.0"),
     ],
@@ -74,6 +77,17 @@ let package = Package(
             name: "FWApplicationCompatible",
             dependencies: ["FWApplication"],
             path: "FWApplication/Classes/Compatible",
+            cSettings: [
+                .define("FWMacroSPM", to: "1")
+            ],
+            swiftSettings: [
+                .define("DEBUG", .when(platforms: [.iOS], configuration: .debug)),
+                .define("FWMacroSPM")
+            ]),
+        .target(
+            name: "FWApplicationSwiftUI",
+            dependencies: ["FWApplicationCompatible"],
+            path: "FWApplication/Classes/SwiftUI",
             cSettings: [
                 .define("FWMacroSPM", to: "1")
             ],